Break tinystan folder into separate files
